requirements.txt
appnope==0.1.0
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
blis==0.4.1
cairocffi==1.2.0
catalogue==1.0.0
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cymem==2.0.4
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
entrypoints==0.3
filelock==3.0.12
future==0.18.2
idna==2.10
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.6.4
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
murmurhash==1.0.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
neo4j==4.1.1
nest-asyncio==1.4.1
nltk==3.5
notebook==6.1.4
numpy==1.19.4
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
plac==1.1.3
preshed==3.0.4
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.1
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-igraph==0.8.3
pytz==2020.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
regex==2020.11.13
requests==2.25.0
sacremoses==0.0.43
selenium==3.141.0
Send2Trash==1.5.0
sentencepiece==0.1.91
six==1.15.0
spacy==2.3.2
srsly==1.0.4
terminado==0.9.1
testpath==0.4.4
texttable==1.6.3
thinc==7.4.1
tokenizers==0.9.3
toml==0.10.2
torch==1.7.0
torchaudio==0.7.0
torchvision==0.8.1
tornado==6.0.4
tqdm==4.52.0
traitlets==5.0.5
transformers==3.5.1
typing-extensions==3.7.4.3
urllib3==1.25.11
wasabi==0.8.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
